Module aws_sdk_elasticbeanstalk::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AbortEnvironmentUpdate

Fluent builder constructing a request to AbortEnvironmentUpdate.

ApplyEnvironmentManagedAction

Fluent builder constructing a request to ApplyEnvironmentManagedAction.

AssociateEnvironmentOperationsRole

Fluent builder constructing a request to AssociateEnvironmentOperationsRole.

CheckDNSAvailability

Fluent builder constructing a request to CheckDNSAvailability.

ComposeEnvironments

Fluent builder constructing a request to ComposeEnvironments.

CreateApplication

Fluent builder constructing a request to CreateApplication.

CreateApplicationVersion

Fluent builder constructing a request to CreateApplicationVersion.

CreateConfigurationTemplate

Fluent builder constructing a request to CreateConfigurationTemplate.

CreateEnvironment

Fluent builder constructing a request to CreateEnvironment.

CreatePlatformVersion

Fluent builder constructing a request to CreatePlatformVersion.

CreateStorageLocation

Fluent builder constructing a request to CreateStorageLocation.

DeleteApplication

Fluent builder constructing a request to DeleteApplication.

DeleteApplicationVersion

Fluent builder constructing a request to DeleteApplicationVersion.

DeleteConfigurationTemplate

Fluent builder constructing a request to DeleteConfigurationTemplate.

DeleteEnvironmentConfiguration

Fluent builder constructing a request to DeleteEnvironmentConfiguration.

DeletePlatformVersion

Fluent builder constructing a request to DeletePlatformVersion.

DescribeAccountAttributes

Fluent builder constructing a request to DescribeAccountAttributes.

DescribeApplicationVersions

Fluent builder constructing a request to DescribeApplicationVersions.

DescribeApplications

Fluent builder constructing a request to DescribeApplications.

DescribeConfigurationOptions

Fluent builder constructing a request to DescribeConfigurationOptions.

DescribeConfigurationSettings

Fluent builder constructing a request to DescribeConfigurationSettings.

DescribeEnvironmentHealth

Fluent builder constructing a request to DescribeEnvironmentHealth.

DescribeEnvironmentManagedActionHistory

Fluent builder constructing a request to DescribeEnvironmentManagedActionHistory.

DescribeEnvironmentManagedActions

Fluent builder constructing a request to DescribeEnvironmentManagedActions.

DescribeEnvironmentResources

Fluent builder constructing a request to DescribeEnvironmentResources.

DescribeEnvironments

Fluent builder constructing a request to DescribeEnvironments.

DescribeEvents

Fluent builder constructing a request to DescribeEvents.

DescribeInstancesHealth

Fluent builder constructing a request to DescribeInstancesHealth.

DescribePlatformVersion

Fluent builder constructing a request to DescribePlatformVersion.

DisassociateEnvironmentOperationsRole

Fluent builder constructing a request to DisassociateEnvironmentOperationsRole.

ListAvailableSolutionStacks

Fluent builder constructing a request to ListAvailableSolutionStacks.

ListPlatformBranches

Fluent builder constructing a request to ListPlatformBranches.

ListPlatformVersions

Fluent builder constructing a request to ListPlatformVersions.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

RebuildEnvironment

Fluent builder constructing a request to RebuildEnvironment.

RequestEnvironmentInfo

Fluent builder constructing a request to RequestEnvironmentInfo.

RestartAppServer

Fluent builder constructing a request to RestartAppServer.

RetrieveEnvironmentInfo

Fluent builder constructing a request to RetrieveEnvironmentInfo.

SwapEnvironmentCNAMEs

Fluent builder constructing a request to SwapEnvironmentCNAMEs.

TerminateEnvironment

Fluent builder constructing a request to TerminateEnvironment.

UpdateApplication

Fluent builder constructing a request to UpdateApplication.

UpdateApplicationResourceLifecycle

Fluent builder constructing a request to UpdateApplicationResourceLifecycle.

UpdateApplicationVersion

Fluent builder constructing a request to UpdateApplicationVersion.

UpdateConfigurationTemplate

Fluent builder constructing a request to UpdateConfigurationTemplate.

UpdateEnvironment

Fluent builder constructing a request to UpdateEnvironment.

UpdateTagsForResource

Fluent builder constructing a request to UpdateTagsForResource.

ValidateConfigurationSettings

Fluent builder constructing a request to ValidateConfigurationSettings.