Struct aws_sdk_elasticbeanstalk::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWSElasticBeanstalkService.

This client allows ergonomic access to a AWSElasticBeanstalkService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn abort_environment_update(&self) -> AbortEnvironmentUpdate<C, M, R>

pub fn apply_environment_managed_action(
&self
) -> ApplyEnvironmentManagedAction<C, M, R>

pub fn associate_environment_operations_role(
&self
) -> AssociateEnvironmentOperationsRole<C, M, R>

pub fn check_dns_availability(&self) -> CheckDNSAvailability<C, M, R>

pub fn compose_environments(&self) -> ComposeEnvironments<C, M, R>

pub fn create_application(&self) -> CreateApplication<C, M, R>

pub fn create_application_version(&self) -> CreateApplicationVersion<C, M, R>

pub fn create_configuration_template(
&self
) -> CreateConfigurationTemplate<C, M, R>

pub fn create_environment(&self) -> CreateEnvironment<C, M, R>

pub fn create_platform_version(&self) -> CreatePlatformVersion<C, M, R>

pub fn create_storage_location(&self) -> CreateStorageLocation<C, M, R>

pub fn delete_application(&self) -> DeleteApplication<C, M, R>

pub fn delete_application_version(&self) -> DeleteApplicationVersion<C, M, R>

pub fn delete_configuration_template(
&self
) -> DeleteConfigurationTemplate<C, M, R>

pub fn delete_environment_configuration(
&self
) -> DeleteEnvironmentConfiguration<C, M, R>

pub fn delete_platform_version(&self) -> DeletePlatformVersion<C, M, R>

pub fn describe_account_attributes(&self) -> DescribeAccountAttributes<C, M, R>

pub fn describe_applications(&self) -> DescribeApplications<C, M, R>

pub fn describe_application_versions(
&self
) -> DescribeApplicationVersions<C, M, R>

pub fn describe_configuration_options(
&self
) -> DescribeConfigurationOptions<C, M, R>

pub fn describe_configuration_settings(
&self
) -> DescribeConfigurationSettings<C, M, R>

pub fn describe_environment_health(&self) -> DescribeEnvironmentHealth<C, M, R>

pub fn describe_environment_managed_action_history(
&self
) -> DescribeEnvironmentManagedActionHistory<C, M, R>

pub fn describe_environment_managed_actions(
&self
) -> DescribeEnvironmentManagedActions<C, M, R>

pub fn describe_environment_resources(
&self
) -> DescribeEnvironmentResources<C, M, R>

pub fn describe_environments(&self) -> DescribeEnvironments<C, M, R>

pub fn describe_events(&self) -> DescribeEvents<C, M, R>

pub fn describe_instances_health(&self) -> DescribeInstancesHealth<C, M, R>

pub fn describe_platform_version(&self) -> DescribePlatformVersion<C, M, R>

pub fn disassociate_environment_operations_role(
&self
) -> DisassociateEnvironmentOperationsRole<C, M, R>

pub fn list_available_solution_stacks(
&self
) -> ListAvailableSolutionStacks<C, M, R>

pub fn list_platform_branches(&self) -> ListPlatformBranches<C, M, R>

pub fn list_platform_versions(&self) -> ListPlatformVersions<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn rebuild_environment(&self) -> RebuildEnvironment<C, M, R>

pub fn request_environment_info(&self) -> RequestEnvironmentInfo<C, M, R>

pub fn restart_app_server(&self) -> RestartAppServer<C, M, R>

pub fn retrieve_environment_info(&self) -> RetrieveEnvironmentInfo<C, M, R>

pub fn swap_environment_cnam_es(&self) -> SwapEnvironmentCNAMEs<C, M, R>

pub fn terminate_environment(&self) -> TerminateEnvironment<C, M, R>

pub fn update_application(&self) -> UpdateApplication<C, M, R>

pub fn update_application_resource_lifecycle(
&self
) -> UpdateApplicationResourceLifecycle<C, M, R>

pub fn update_application_version(&self) -> UpdateApplicationVersion<C, M, R>

pub fn update_configuration_template(
&self
) -> UpdateConfigurationTemplate<C, M, R>

pub fn update_environment(&self) -> UpdateEnvironment<C, M, R>

pub fn update_tags_for_resource(&self) -> UpdateTagsForResource<C, M, R>

pub fn validate_configuration_settings(
&self
) -> ValidateConfigurationSettings<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self