Module aws_sdk_elasticbeanstalk::error

Modules

code_build_not_in_service_region_exception

See CodeBuildNotInServiceRegionException

elastic_beanstalk_service_exception

See ElasticBeanstalkServiceException

insufficient_privileges_exception

See InsufficientPrivilegesException

invalid_request_exception

See InvalidRequestException

managed_action_invalid_state_exception

See ManagedActionInvalidStateException

operation_in_progress_exception

See OperationInProgressException

platform_version_still_referenced_exception

See PlatformVersionStillReferencedException

resource_not_found_exception

See ResourceNotFoundException

resource_type_not_supported_exception

See ResourceTypeNotSupportedException

s3_location_not_in_service_region_exception

See S3LocationNotInServiceRegionException

s3_subscription_required_exception

See S3SubscriptionRequiredException

source_bundle_deletion_exception

See SourceBundleDeletionException

too_many_application_versions_exception

See TooManyApplicationVersionsException

too_many_applications_exception

See TooManyApplicationsException

too_many_buckets_exception

See TooManyBucketsException

too_many_configuration_templates_exception

See TooManyConfigurationTemplatesException

too_many_environments_exception

See TooManyEnvironmentsException

too_many_platforms_exception

See TooManyPlatformsException

too_many_tags_exception

See TooManyTagsException

Structs

AbortEnvironmentUpdateError

ApplyEnvironmentManagedActionError

AssociateEnvironmentOperationsRoleError

CheckDNSAvailabilityError

CodeBuildNotInServiceRegionException

AWS CodeBuild is not available in the specified region.

ComposeEnvironmentsError

CreateApplicationError

CreateApplicationVersionError

CreateConfigurationTemplateError

CreateEnvironmentError

CreatePlatformVersionError

CreateStorageLocationError

DeleteApplicationError

DeleteApplicationVersionError

DeleteConfigurationTemplateError

DeleteEnvironmentConfigurationError

DeletePlatformVersionError

DescribeAccountAttributesError

DescribeApplicationVersionsError

DescribeApplicationsError

DescribeConfigurationOptionsError

DescribeConfigurationSettingsError

DescribeEnvironmentHealthError

DescribeEnvironmentManagedActionHistoryError

DescribeEnvironmentManagedActionsError

DescribeEnvironmentResourcesError

DescribeEnvironmentsError

DescribeEventsError

DescribeInstancesHealthError

DescribePlatformVersionError

DisassociateEnvironmentOperationsRoleError

ElasticBeanstalkServiceException

A generic service exception has occurred.

InsufficientPrivilegesException

The specified account does not have sufficient privileges for one or more AWS

InvalidRequestException

One or more input parameters is not valid. Please correct the input parameters and try

ListAvailableSolutionStacksError

ListPlatformBranchesError

ListPlatformVersionsError

ListTagsForResourceError

ManagedActionInvalidStateException

Cannot modify the managed action in its current state.

OperationInProgressException

Unable to perform the specified operation because another operation that effects an

PlatformVersionStillReferencedException

You cannot delete the platform version because there are still environments running on it.

RebuildEnvironmentError

RequestEnvironmentInfoError

ResourceNotFoundException

A resource doesn't exist for the specified Amazon Resource Name (ARN).

ResourceTypeNotSupportedException

The type of the specified Amazon Resource Name (ARN) isn't supported for this operation.

RestartAppServerError

RetrieveEnvironmentInfoError

S3LocationNotInServiceRegionException

The specified S3 bucket does not belong to the S3 region in which the service is

S3SubscriptionRequiredException

The specified account does not have a subscription to Amazon S3.

SourceBundleDeletionException

Unable to delete the Amazon S3 source bundle associated with the application version.

SwapEnvironmentCNAMEsError

TerminateEnvironmentError

TooManyApplicationVersionsException

The specified account has reached its limit of application versions.

TooManyApplicationsException

The specified account has reached its limit of applications.

TooManyBucketsException

The specified account has reached its limit of Amazon S3 buckets.

TooManyConfigurationTemplatesException

The specified account has reached its limit of configuration templates.

TooManyEnvironmentsException

The specified account has reached its limit of environments.

TooManyPlatformsException

You have exceeded the maximum number of allowed platforms associated with the account.

TooManyTagsException

The number of tags in the resource would exceed the number of tags that each resource

UpdateApplicationError

UpdateApplicationResourceLifecycleError

UpdateApplicationVersionError

UpdateConfigurationTemplateError

UpdateEnvironmentError

UpdateTagsForResourceError

ValidateConfigurationSettingsError

Enums

AbortEnvironmentUpdateErrorKind

ApplyEnvironmentManagedActionErrorKind

AssociateEnvironmentOperationsRoleErrorKind

CheckDNSAvailabilityErrorKind

ComposeEnvironmentsErrorKind

CreateApplicationErrorKind

CreateApplicationVersionErrorKind

CreateConfigurationTemplateErrorKind

CreateEnvironmentErrorKind

CreatePlatformVersionErrorKind

CreateStorageLocationErrorKind

DeleteApplicationErrorKind

DeleteApplicationVersionErrorKind

DeleteConfigurationTemplateErrorKind

DeleteEnvironmentConfigurationErrorKind

DeletePlatformVersionErrorKind

DescribeAccountAttributesErrorKind

DescribeApplicationVersionsErrorKind

DescribeApplicationsErrorKind

DescribeConfigurationOptionsErrorKind

DescribeConfigurationSettingsErrorKind

DescribeEnvironmentHealthErrorKind

DescribeEnvironmentManagedActionHistoryErrorKind

DescribeEnvironmentManagedActionsErrorKind

DescribeEnvironmentResourcesErrorKind

DescribeEnvironmentsErrorKind

DescribeEventsErrorKind

DescribeInstancesHealthErrorKind

DescribePlatformVersionErrorKind

DisassociateEnvironmentOperationsRoleErrorKind

ListAvailableSolutionStacksErrorKind

ListPlatformBranchesErrorKind

ListPlatformVersionsErrorKind

ListTagsForResourceErrorKind

RebuildEnvironmentErrorKind

RequestEnvironmentInfoErrorKind

RestartAppServerErrorKind

RetrieveEnvironmentInfoErrorKind

SwapEnvironmentCNAMEsErrorKind

TerminateEnvironmentErrorKind

UpdateApplicationErrorKind

UpdateApplicationResourceLifecycleErrorKind

UpdateApplicationVersionErrorKind

UpdateConfigurationTemplateErrorKind

UpdateEnvironmentErrorKind

UpdateTagsForResourceErrorKind

ValidateConfigurationSettingsErrorKind