Module aws_sdk_elasticbeanstalk::input

Modules

abort_environment_update_input

See AbortEnvironmentUpdateInput

apply_environment_managed_action_input

See ApplyEnvironmentManagedActionInput

associate_environment_operations_role_input

See AssociateEnvironmentOperationsRoleInput

check_dns_availability_input

See CheckDnsAvailabilityInput

compose_environments_input

See ComposeEnvironmentsInput

create_application_input

See CreateApplicationInput

create_application_version_input

See CreateApplicationVersionInput

create_configuration_template_input

See CreateConfigurationTemplateInput

create_environment_input

See CreateEnvironmentInput

create_platform_version_input

See CreatePlatformVersionInput

create_storage_location_input

See CreateStorageLocationInput

delete_application_input

See DeleteApplicationInput

delete_application_version_input

See DeleteApplicationVersionInput

delete_configuration_template_input

See DeleteConfigurationTemplateInput

delete_environment_configuration_input

See DeleteEnvironmentConfigurationInput

delete_platform_version_input

See DeletePlatformVersionInput

describe_account_attributes_input

See DescribeAccountAttributesInput

describe_application_versions_input

See DescribeApplicationVersionsInput

describe_applications_input

See DescribeApplicationsInput

describe_configuration_options_input

See DescribeConfigurationOptionsInput

describe_configuration_settings_input

See DescribeConfigurationSettingsInput

describe_environment_health_input

See DescribeEnvironmentHealthInput

describe_environment_managed_action_history_input

See DescribeEnvironmentManagedActionHistoryInput

describe_environment_managed_actions_input

See DescribeEnvironmentManagedActionsInput

describe_environment_resources_input

See DescribeEnvironmentResourcesInput

describe_environments_input

See DescribeEnvironmentsInput

describe_events_input

See DescribeEventsInput

describe_instances_health_input

See DescribeInstancesHealthInput

describe_platform_version_input

See DescribePlatformVersionInput

disassociate_environment_operations_role_input

See DisassociateEnvironmentOperationsRoleInput

list_available_solution_stacks_input

See ListAvailableSolutionStacksInput

list_platform_branches_input

See ListPlatformBranchesInput

list_platform_versions_input

See ListPlatformVersionsInput

list_tags_for_resource_input

See ListTagsForResourceInput

rebuild_environment_input

See RebuildEnvironmentInput

request_environment_info_input

See RequestEnvironmentInfoInput

restart_app_server_input

See RestartAppServerInput

retrieve_environment_info_input

See RetrieveEnvironmentInfoInput

swap_environment_cnam_es_input

See SwapEnvironmentCnamEsInput

terminate_environment_input

See TerminateEnvironmentInput

update_application_input

See UpdateApplicationInput

update_application_resource_lifecycle_input

See UpdateApplicationResourceLifecycleInput

update_application_version_input

See UpdateApplicationVersionInput

update_configuration_template_input

See UpdateConfigurationTemplateInput

update_environment_input

See UpdateEnvironmentInput

update_tags_for_resource_input

See UpdateTagsForResourceInput

validate_configuration_settings_input

See ValidateConfigurationSettingsInput

Structs

AbortEnvironmentUpdateInput

ApplyEnvironmentManagedActionInput

Request to execute a scheduled managed action immediately.

AssociateEnvironmentOperationsRoleInput

Request to add or change the operations role used by an environment.

CheckDnsAvailabilityInput

Results message indicating whether a CNAME is available.

ComposeEnvironmentsInput

Request to create or update a group of environments.

CreateApplicationInput

Request to create an application.

CreateApplicationVersionInput

CreateConfigurationTemplateInput

Request to create a configuration template.

CreateEnvironmentInput

CreatePlatformVersionInput

Request to create a new platform version.

CreateStorageLocationInput

DeleteApplicationInput

Request to delete an application.

DeleteApplicationVersionInput

Request to delete an application version.

DeleteConfigurationTemplateInput

Request to delete a configuration template.

DeleteEnvironmentConfigurationInput

Request to delete a draft environment configuration.

DeletePlatformVersionInput

DescribeAccountAttributesInput

DescribeApplicationVersionsInput

Request to describe application versions.

DescribeApplicationsInput

Request to describe one or more applications.

DescribeConfigurationOptionsInput

Result message containing a list of application version descriptions.

DescribeConfigurationSettingsInput

Result message containing all of the configuration settings for a specified solution

DescribeEnvironmentHealthInput

See the example below to learn how to create a request body.

DescribeEnvironmentManagedActionHistoryInput

Request to list completed and failed managed actions.

DescribeEnvironmentManagedActionsInput

Request to list an environment's upcoming and in-progress managed actions.

DescribeEnvironmentResourcesInput

Request to describe the resources in an environment.

DescribeEnvironmentsInput

Request to describe one or more environments.

DescribeEventsInput

Request to retrieve a list of events for an environment.

DescribeInstancesHealthInput

Parameters for a call to DescribeInstancesHealth.

DescribePlatformVersionInput

DisassociateEnvironmentOperationsRoleInput

Request to disassociate the operations role from an environment.

ListAvailableSolutionStacksInput

ListPlatformBranchesInput

ListPlatformVersionsInput

ListTagsForResourceInput

RebuildEnvironmentInput

RequestEnvironmentInfoInput

Request to retrieve logs from an environment and store them in your Elastic Beanstalk

RestartAppServerInput

RetrieveEnvironmentInfoInput

Request to download logs retrieved with RequestEnvironmentInfo.

SwapEnvironmentCnamEsInput

Swaps the CNAMEs of two environments.

TerminateEnvironmentInput

Request to terminate an environment.

UpdateApplicationInput

Request to update an application.

UpdateApplicationResourceLifecycleInput

UpdateApplicationVersionInput

UpdateConfigurationTemplateInput

The result message containing the options for the specified solution stack.

UpdateEnvironmentInput

Request to update an environment.

UpdateTagsForResourceInput

ValidateConfigurationSettingsInput

A list of validation messages for a specified configuration template.