Module aws_sdk_elasticbeanstalk::model

Modules

application_description

See ApplicationDescription

application_metrics

See ApplicationMetrics

application_resource_lifecycle_config

See ApplicationResourceLifecycleConfig

application_version_description

See ApplicationVersionDescription

application_version_lifecycle_config

See ApplicationVersionLifecycleConfig

auto_scaling_group

See AutoScalingGroup

build_configuration

See BuildConfiguration

builder

See Builder

configuration_option_description

See ConfigurationOptionDescription

configuration_option_setting

See ConfigurationOptionSetting

configuration_settings_description

See ConfigurationSettingsDescription

cpu_utilization

See CpuUtilization

custom_ami

See CustomAmi

deployment

See Deployment

environment_description

See EnvironmentDescription

environment_info_description

See EnvironmentInfoDescription

environment_link

See EnvironmentLink

environment_resource_description

See EnvironmentResourceDescription

environment_resources_description

See EnvironmentResourcesDescription

environment_tier

See EnvironmentTier

event_description

See EventDescription

instance

See Instance

instance_health_summary

See InstanceHealthSummary

latency

See Latency

launch_configuration

See LaunchConfiguration

launch_template

See LaunchTemplate

listener

See Listener

load_balancer

See LoadBalancer

load_balancer_description

See LoadBalancerDescription

managed_action

See ManagedAction

managed_action_history_item

See ManagedActionHistoryItem

max_age_rule

See MaxAgeRule

max_count_rule

See MaxCountRule

option_restriction_regex

See OptionRestrictionRegex

option_specification

See OptionSpecification

platform_branch_summary

See PlatformBranchSummary

platform_description

See PlatformDescription

platform_filter

See PlatformFilter

platform_framework

See PlatformFramework

platform_programming_language

See PlatformProgrammingLanguage

platform_summary

See PlatformSummary

queue

See Queue

resource_quota

See ResourceQuota

resource_quotas

See ResourceQuotas

s3_location

See S3Location

search_filter

See SearchFilter

single_instance_health

See SingleInstanceHealth

solution_stack_description

See SolutionStackDescription

source_build_information

See SourceBuildInformation

source_configuration

See SourceConfiguration

status_codes

See StatusCodes

system_status

See SystemStatus

tag

See Tag

trigger

See Trigger

validation_message

See ValidationMessage

Structs

ApplicationDescription

Describes the properties of an application.

ApplicationMetrics

Application request metrics for an AWS Elastic Beanstalk environment.

ApplicationResourceLifecycleConfig

The resource lifecycle configuration for an application. Defines lifecycle settings for

ApplicationVersionDescription

Describes the properties of an application version.

ApplicationVersionLifecycleConfig

The application version lifecycle settings for an application. Defines the rules that

AutoScalingGroup

Describes an Auto Scaling launch configuration.

BuildConfiguration

Settings for an AWS CodeBuild build.

Builder

The builder used to build the custom platform.

ConfigurationOptionDescription

Describes the possible values for a configuration option.

ConfigurationOptionSetting

A specification identifying an individual configuration option along with its current

ConfigurationSettingsDescription

Describes the settings for a configuration set.

CpuUtilization

CPU utilization metrics for an instance.

CustomAmi

A custom AMI available to platforms.

Deployment

Information about an application version deployment.

EnvironmentDescription

Describes the properties of an environment.

EnvironmentInfoDescription

The information retrieved from the Amazon EC2 instances.

EnvironmentLink

A link to another environment, defined in the environment's manifest. Links provide

EnvironmentResourceDescription

Describes the AWS resources in use by this environment. This data is live.

EnvironmentResourcesDescription

Describes the AWS resources in use by this environment. This data is not live

EnvironmentTier

Describes the properties of an environment tier

EventDescription

Describes an event.

Instance

The description of an Amazon EC2 instance.

InstanceHealthSummary

Represents summary information about the health of an instance. For more information,

Latency

Represents the average latency for the slowest X percent of requests over the last 10

LaunchConfiguration

Describes an Auto Scaling launch configuration.

LaunchTemplate

Describes an Amazon EC2 launch template.

Listener

Describes the properties of a Listener for the LoadBalancer.

LoadBalancer

Describes a LoadBalancer.

LoadBalancerDescription

Describes the details of a LoadBalancer.

ManagedAction

The record of an upcoming or in-progress managed action.

ManagedActionHistoryItem

The record of a completed or failed managed action.

MaxAgeRule

A lifecycle rule that deletes application versions after the specified number of

MaxCountRule

A lifecycle rule that deletes the oldest application version when the maximum count is

OptionRestrictionRegex

A regular expression representing a restriction on a string configuration option

OptionSpecification

A specification identifying an individual configuration option.

PlatformBranchSummary

Summary information about a platform branch.

PlatformDescription

Detailed information about a platform version.

PlatformFilter

Describes criteria to restrict the results when listing platform versions.

PlatformFramework

A framework supported by the platform.

PlatformProgrammingLanguage

A programming language supported by the platform.

PlatformSummary

Summary information about a platform version.

Queue

Describes a queue.

ResourceQuota

The AWS Elastic Beanstalk quota information for a single resource type in an AWS account. It

ResourceQuotas

A set of per-resource AWS Elastic Beanstalk quotas associated with an AWS account. They reflect

S3Location

The bucket and key of an item stored in Amazon S3.

SearchFilter

Describes criteria to restrict a list of results.

SingleInstanceHealth

Detailed health information about an Amazon EC2 instance in your Elastic Beanstalk

SolutionStackDescription

Describes the solution stack.

SourceBuildInformation

Location of the source code for an application version.

SourceConfiguration

A specification for an environment configuration.

StatusCodes

Represents the percentage of requests over the last 10 seconds that resulted in each

SystemStatus

CPU utilization and load average metrics for an Amazon EC2 instance.

Tag

Describes a tag applied to a resource in an environment.

Trigger

Describes a trigger.

ValidationMessage

An error or warning for a desired configuration option value.

Enums

ActionHistoryStatus

NOTE: ActionHistoryStatus::Unknown has been renamed to ::UnknownValue.

ActionStatus

NOTE: ActionStatus::Unknown has been renamed to ::UnknownValue.

ActionType

NOTE: ActionType::Unknown has been renamed to ::UnknownValue.

ApplicationVersionStatus

ComputeType

ConfigurationDeploymentStatus

ConfigurationOptionValueType

EnvironmentHealth

EnvironmentHealthAttribute

EnvironmentHealthStatus

NOTE: EnvironmentHealthStatus::Unknown has been renamed to ::UnknownValue.

EnvironmentInfoType

EnvironmentStatus

EventSeverity

FailureType

InstancesHealthAttribute

PlatformStatus

SourceRepository

SourceType

ValidationSeverity