Module aws_sdk_elasticbeanstalk::operation

Structs

AbortEnvironmentUpdate

Cancels in-progress environment configuration update or application version

ApplyEnvironmentManagedAction

Applies a scheduled managed action immediately. A managed action can be applied only if

AssociateEnvironmentOperationsRole

Add or change the operations role used by an environment. After this call is made, Elastic Beanstalk

CheckDNSAvailability

Checks if the specified CNAME is available.

ComposeEnvironments

Create or update a group of environments that each run a separate component of a single

CreateApplication

Creates an application that has one configuration template named default

CreateApplicationVersion

Creates an application version for the specified application. You can create an

CreateConfigurationTemplate

Creates an AWS Elastic Beanstalk configuration template, associated with a specific Elastic Beanstalk

CreateEnvironment

Launches an AWS Elastic Beanstalk environment for the specified application using the specified

CreatePlatformVersion

Create a new version of your custom platform.

CreateStorageLocation

Creates a bucket in Amazon S3 to store application versions, logs, and other files used

DeleteApplication

Deletes the specified application along with all associated versions and

DeleteApplicationVersion

Deletes the specified version from the specified application.

DeleteConfigurationTemplate

Deletes the specified configuration template.

DeleteEnvironmentConfiguration

Deletes the draft configuration associated with the running environment.

DeletePlatformVersion

Deletes the specified version of a custom platform.

DescribeAccountAttributes

Returns attributes related to AWS Elastic Beanstalk that are associated with the calling AWS

DescribeApplicationVersions

Retrieve a list of application versions.

DescribeApplications

Returns the descriptions of existing applications.

DescribeConfigurationOptions

Describes the configuration options that are used in a particular configuration

DescribeConfigurationSettings

Returns a description of the settings for the specified configuration set, that is,

DescribeEnvironmentHealth

Returns information about the overall health of the specified environment. The

DescribeEnvironmentManagedActionHistory

Lists an environment's completed and failed managed actions.

DescribeEnvironmentManagedActions

Lists an environment's upcoming and in-progress managed actions.

DescribeEnvironmentResources

Returns AWS resources for this environment.

DescribeEnvironments

Returns descriptions for existing environments.

DescribeEvents

Returns list of event descriptions matching criteria up to the last 6 weeks.

DescribeInstancesHealth

Retrieves detailed information about the health of instances in your AWS Elastic

DescribePlatformVersion

Describes a platform version. Provides full details. Compare to ListPlatformVersions, which provides summary information about a list of

DisassociateEnvironmentOperationsRole

Disassociate the operations role from an environment. After this call is made, Elastic Beanstalk uses

ListAvailableSolutionStacks

Returns a list of the available solution stack names, with the public version first and

ListPlatformBranches

Lists the platform branches available for your account in an AWS Region. Provides

ListPlatformVersions

Lists the platform versions available for your account in an AWS Region. Provides

ListTagsForResource

Return the tags applied to an AWS Elastic Beanstalk resource. The response contains a list of tag key-value pairs.

RebuildEnvironment

Deletes and recreates all of the AWS resources (for example: the Auto Scaling group,

RequestEnvironmentInfo

Initiates a request to compile the specified type of information of the deployed

RestartAppServer

Causes the environment to restart the application container server running on each

RetrieveEnvironmentInfo

Retrieves the compiled information from a RequestEnvironmentInfo

SwapEnvironmentCNAMEs

Swaps the CNAMEs of two environments.

TerminateEnvironment

Terminates the specified environment.

UpdateApplication

Updates the specified application to have the specified properties.

UpdateApplicationResourceLifecycle

Modifies lifecycle settings for an application.

UpdateApplicationVersion

Updates the specified application version to have the specified properties.

UpdateConfigurationTemplate

Updates the specified configuration template to have the specified properties or

UpdateEnvironment

Updates the environment description, deploys a new application version, updates the

UpdateTagsForResource

Update the list of tags applied to an AWS Elastic Beanstalk resource. Two lists can be passed: TagsToAdd

ValidateConfigurationSettings

Takes a set of configuration settings and either a configuration template or