Struct aws_sdk_elasticbeanstalk::output::update_environment_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for UpdateEnvironmentOutput

Implementations

impl Builder

pub fn environment_name(self, input: impl Into<String>) -> Self

The name of this environment.

pub fn set_environment_name(self, input: Option<String>) -> Self

pub fn environment_id(self, input: impl Into<String>) -> Self

The ID of this environment.

pub fn set_environment_id(self, input: Option<String>) -> Self

pub fn application_name(self, input: impl Into<String>) -> Self

The name of the application associated with this environment.

pub fn set_application_name(self, input: Option<String>) -> Self

pub fn version_label(self, input: impl Into<String>) -> Self

The application version deployed in this environment.

pub fn set_version_label(self, input: Option<String>) -> Self

pub fn solution_stack_name(self, input: impl Into<String>) -> Self

The name of the SolutionStack deployed with this environment.

pub fn set_solution_stack_name(self, input: Option<String>) -> Self

pub fn platform_arn(self, input: impl Into<String>) -> Self

The ARN of the platform version.

pub fn set_platform_arn(self, input: Option<String>) -> Self

pub fn template_name(self, input: impl Into<String>) -> Self

The name of the configuration template used to originally launch this environment.

pub fn set_template_name(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

Describes this environment.

pub fn set_description(self, input: Option<String>) -> Self

pub fn endpoint_url(self, input: impl Into<String>) -> Self

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

pub fn set_endpoint_url(self, input: Option<String>) -> Self

pub fn cname(self, input: impl Into<String>) -> Self

The URL to the CNAME for this environment.

pub fn set_cname(self, input: Option<String>) -> Self

pub fn date_created(self, input: Instant) -> Self

The creation date for this environment.

pub fn set_date_created(self, input: Option<Instant>) -> Self

pub fn date_updated(self, input: Instant) -> Self

The last modified date for this environment.

pub fn set_date_updated(self, input: Option<Instant>) -> Self

pub fn status(self, input: EnvironmentStatus) -> Self

The current operational status of the environment:

Launching: Environment is in the process of initial deployment.

Updating: Environment is in the process of updating its configuration settings or application version.

Ready: Environment is available to have an action performed on it, such as update or terminate.

Terminating: Environment is in the shut-down process.

Terminated: Environment is not running.

pub fn set_status(self, input: Option<EnvironmentStatus>) -> Self

pub fn abortable_operation_in_progress(self, input: bool) -> Self

Indicates if there is an in-progress environment configuration update or application version deployment that you can cancel.

true: There is an update in progress.

false: There are no updates currently in progress.

pub fn set_abortable_operation_in_progress(self, input: Option<bool>) -> Self

pub fn health(self, input: EnvironmentHealth) -> Self

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment:

Red: Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment.

Yellow: Indicates that something is wrong. Occurs when two consecutive failures occur for an environment.

Green: Indicates the environment is healthy and fully functional.

Grey: Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironment request.

Default: Grey

pub fn set_health(self, input: Option<EnvironmentHealth>) -> Self

pub fn health_status(self, input: EnvironmentHealthStatus) -> Self

Returns the health status of the application running in your environment. For more information, see Health Colors and Statuses.

pub fn set_health_status(self, input: Option<EnvironmentHealthStatus>) -> Self

pub fn resources(self, input: EnvironmentResourcesDescription) -> Self

The description of the AWS resources used by this environment.

pub fn set_resources(
self,
input: Option<EnvironmentResourcesDescription>
) -> Self

pub fn tier(self, input: EnvironmentTier) -> Self

Describes the current tier of this environment.

pub fn set_tier(self, input: Option<EnvironmentTier>) -> Self

pub fn environment_links(self, input: impl Into<EnvironmentLink>) -> Self

pub fn set_environment_links(self, input: Option<Vec<EnvironmentLink>>) -> Self

pub fn environment_arn(self, input: impl Into<String>) -> Self

The environment's Amazon Resource Name (ARN), which can be used in other API requests that require an ARN.

pub fn set_environment_arn(self, input: Option<String>) -> Self

pub fn operations_role(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the environment's operations role. For more information, see Operations roles in the AWS Elastic Beanstalk Developer Guide.

pub fn set_operations_role(self, input: Option<String>) -> Self

pub fn build(self) -> UpdateEnvironmentOutput

Consumes the builder and constructs a UpdateEnvironmentOutput